Fault tolerance improvement

+++ controllers/drivers/info_drivers.php
@@ -19,7 +19,7 @@ if(isset($_GET["url"])) {
         $array_value[] = $value;
     }
 
-    if (count($array_value) != 0)
+    if (count($array_value) != 0 && file_exists(BACKUP_FILE))
         file_put_contents($backupFile, json_encode($array_value));
     else
         $array_value = json_decode(file_get_contents($backupFile));
